add celery + unit tests

+++ dashboard/apis.py
@@ -1,16 +1,13 @@
 import logging
 
 from django.contrib.auth.decorators import login_required
-from django.http import JsonResponse, HttpResponse
-from django.template import defaultfilters
-from django.utils import timezone
-from django.urls import reverse
+from django.http import JsonResponse
 
 from websites.utils import is_ajax, explode_airbnb_url
-from websites.config import WEBSITE_URL, MAX_WEBSITES_COUNT
+from websites.config import MAX_WEBSITES_COUNT
 from websites.models import Website
 
-from scrapper.views import scrap_airbnb_data
+from .tasks import scrap_and_create_website
 
 
 _logger = logging.getLogger('websites')
@@ -39,25 +36,14 @@ def api_website_create(request):
         return _user_error("L'URL fournie n'est pas une URL Airbnb valide")
 
     # check if the current user does not exceed limitations
-    if Website.objects.filter(user=request.user).count() >= MAX_WEBSITES_COUNT:
+    if Website.has_reached_resource_limits(request.user):
         return _user_error("Limite de nombre de sites atteinte. Supprimez-en un pour pouvoir en créer un nouveau.")
 
-    # send the URL to the scrapper
-    try:
-        data = scrap_airbnb_data(airbnb_id)
-    except Exception:
-        return _internal_error("Impossible d'accéder à votre annonce Airbnb")
+    # scrap and create the website in background.
+    # the task id is provided to allow the client to poll task result
+    task = scrap_and_create_website.delay(request.user, rental_base_url, airbnb_id)
 
-    # generate the website and get the redirect page
-    website = Website.create(request.user, rental_base_url, data)
-
-    return JsonResponse({
-        "key": website.key,
-        "name": website.name,
-        "url": WEBSITE_URL % website.key,
-        "generated_date": defaultfilters.date(timezone.localtime(website.generated_date), "d/m/Y G:i"),
-        "delete_url": reverse('api_website_delete', args=[website.key]),
-    })
+    return JsonResponse({"task_id": task.id}, status=202)
 
 
 @login_required
@@ -67,10 +53,10 @@ def api_website_delete(request, key):
     if not is_ajax(request):
         return _internal_error("mauvaise requête")
 
-    try:
-        website = Website.objects.get(key=key)
-    except Website.DoesNotExist:
-        return _user_error("Ce site n'existe pas")
+    website = Website.get_website(key)
 
-    website.delete()
-    return JsonResponse({"key": key})
+    if website:
+        website.delete()
+        return JsonResponse({"key": key})
+    else:
+        return _user_error("Ce site n'existe pas")
